requirements.txt
faiss-cpu
img2vec_pytorch
torch
torchvision
numpy
tqdm
clint